requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi>=1.17.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==3.0.2
defusedxml==0.7.1
dill==0.3.8
dj-rest-auth==6.0.0
Django==4.2.14
django-allauth==0.61.1
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-ratelimit==4.1.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
django-timezone-field==7.0
django-types==0.19.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.1
execnet==2.1.1
factory-boy==3.3.0
Faker==26.0.0
filelock==3.16.1
firebase-admin==6.6.0
fsspec==2024.10.0
gevent==24.11.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==22.0.0
hiredis==2.3.2
httplib2==0.22.0
huggingface-hub==0.26.3
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
mccabe==0.7.0
msgpack==1.0.8
numpy==2.1.3
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.2.0
pytest==8.2.2
pytest-django==4.8.0
pytest-xdist==3.6.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.0
rsa==4.9
ruff==0.4.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sendgrid==6.11.0
service-identity==24.1.0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.20.3
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.67.1
transformers==4.46.0
Twisted==24.3.0
txaio==23.1.1
types-psycopg2==2.9.21.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
django-admin-interface==0.29.2
django-pwa-webpush==1.1.3
django-templated-email==3.0.1
pytest-mock==3.14.0
tabulate==0.9.0
django-filter==24.3
pytest-asyncio==0.25.3
social-auth-app-django==5.4.3
django-otp==1.5.4